% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp2vis.R
\name{data_filt}
\alias{data_filt}
\title{Function filters data from missing values and expression lower than certain threshold; Functions works on the output of read functions as exp_data_input and sample_data_input}
\usage{
data_filt()
}
\description{
Function filters data from missing values and expression lower than certain threshold; Functions works on the output of read functions as exp_data_input and sample_data_input
}
