% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp2vis.R
\name{deseq_analysis}
\alias{deseq_analysis}
\title{Function define two groups, perform analysis using DESeq2 library and visulaize data with set padj and log fold change as a Volcano plot}
\usage{
deseq_analysis()
}
\description{
Function define two groups, perform analysis using DESeq2 library and visulaize data with set padj and log fold change as a Volcano plot
}
