% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp2vis.R
\name{read_external_file}
\alias{read_external_file}
\title{Function check for library dependencies (first time) then reads user files as expression data and sample phenotype data as tab delimited files and extract needed data for further analysis; format inputs for data_filt(), deseq_functiona nd vis-function}
\usage{
read_external_file()
}
\description{
Function check for library dependencies (first time) then reads user files as expression data and sample phenotype data as tab delimited files and extract needed data for further analysis; format inputs for data_filt(), deseq_functiona nd vis-function
}
