% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp2vis.R
\name{read_gse}
\alias{read_gse}
\title{Function check for library dependencies (first time) then reads GEO GSE ID; specifies platform -GPL- extracts expression and pheotype data then format data inputs for data_filt()m deseq_function(),and vis_function}
\usage{
read_gse(gse_id)
}
\arguments{
\item{gse_id}{character of GEO GSE ID}
}
\description{
Function check for library dependencies (first time) then reads GEO GSE ID; specifies platform -GPL- extracts expression and pheotype data then format data inputs for data_filt()m deseq_function(),and vis_function
}
