% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp2vis.R
\name{read_series_matrix}
\alias{read_series_matrix}
\title{Function check for library dependencies (first time) then reads GEO GSE series matrix zipped file; extracts expression and pheotype data then format data inputs for data_filt()m deseq_function(),and vis_function}
\usage{
read_series_matrix()
}
\description{
Function check for library dependencies (first time) then reads GEO GSE series matrix zipped file; extracts expression and pheotype data then format data inputs for data_filt()m deseq_function(),and vis_function
}
