% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp2vis.R
\name{vis_function}
\alias{vis_function}
\title{Function set number of Up- and down- regulated genes and build visual analysis: Pearson's correlation matrix for each group; Clustered heatmap and annotated to 20 pathways in both Up and Down regulated}
\usage{
vis_function()
}
\description{
Function set number of Up- and down- regulated genes and build visual analysis: Pearson's correlation matrix for each group; Clustered heatmap and annotated to 20 pathways in both Up and Down regulated
}
